+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-03-09
+### Added
+- Add certificate manipulations.
+- Added serial number and uids to cert config.
+
+### Changed
+- Added changelog.
+- Added some small benchmarks.
+- Minor cert example doc change.
+
 ## [0.1.0] - 2023-03-04
 ### Added
 - Admission extension now works.
@@ -57,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add github workflow
 - initial commit
 
-[Unreleased]: https://github.com/wokdav/gopki/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/wokdav/gopki/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/wokdav/gopki/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/wokdav/gopki/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/wokdav/gopki/releases/tag/v0.0.1